Rename query file
