+++ components/dashboard/src/data/organizations/org-workspace-classes-query.ts
